requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
Deprecated==1.2.13
Django==4.0
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-debug-toolbar==3.6.0
django-extensions==3.2.0
django-timezone-field==5.0
flower==1.2.0
hedera-sdk-py==2.17.3
humanize==4.3.0
idna==3.3
JPype1==1.4.0
kombu==5.2.4
packaging==21.3
Pillow==9.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
psycopg2-binary==2.9.3
pyjnius==1.4.2
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tornado==6.2
tzdata==2022.2
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
